risk-assessment-workflow/requirements.txt
# Azure AI Foundry Projects SDK (for hosted agent invocation)
azure-ai-projects>=2.0.0

# Azure Identity
azure-identity>=1.15.0

# Data models
pydantic>=2.0.0

# Environment
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
